When deploying subctl, fall back to devel
During releases from branches, the e2e jobs look for a subctl matching
the target release; but that doesn't exist yet, so the download fails.
To avoid that, if installing the target release version fails, fall
back to devel subctl (which should provide the behaviour we need for a
release, and is more appropriate than the default, which is the latest
release).

Signed-off-by: Stephen Kitt <skitt@redhat.com>